refactor: move logic into dedicated files
