+++ integrations/wechat/src/setup.ts
@@ -1,0 +1,4 @@
+import * as bp from '.botpress'
+
+export const register: bp.IntegrationProps['register'] = async (props) => {}
+export const unregister: bp.IntegrationProps['register'] = async (props) => {}
